Output JS bundle to dist directly
`webpack.client.ts` was configured to always output the bundle file to
the `src/server/public/js` folder, even when building the production
bundle. This is not good as it means that building the production bundle
while the command `watch-client` is executing would overwrite the
bundle generated by the latter.

I changed the `webpack.client.ts` to check whether it is being executed
in production or development mode. If it is the former, then the output
is automatically written under `dist/public/js`.